промежуточный коммит

+++ src/api/v1/users.py
@@ -1,10 +1,68 @@
-from fastapi import APIRouter, Depends
+from functools import lru_cache, wraps
+
+from fastapi import APIRouter, Depends, Cookie
 from fastapi.responses import JSONResponse, Response
+from jose import JWTError, jwt
 
 from api.v1.models.users.results.user_result import UserResult
+from configs.security_settings import SecuritySettings
 from services.user_service import UserService, get_user_service
 
 router = APIRouter()
+
+
+@lru_cache()
+def get_settings():
+    return SecuritySettings()
+
+
+def encode_jwt():
+    def wrapper(func):
+        @wraps(func)
+        def inner(*args, **kwargs):
+            result = func(*args, **kwargs)
+            if result and isinstance(result, dict):
+                return jwt.encode(result, get_settings().openssl_key, algorithm=get_settings().algorithm)
+            else:
+                return result
+
+        return inner
+
+    return wrapper
+
+
+def decode_jwt():
+    def wrapper(func):
+        @wraps(func)
+        def inner(*args, **kwargs):
+            try:
+                token = func(*args, **kwargs)
+                return jwt.decode(token, get_settings().openssl_key, algorithms=[get_settings().algorithm])
+            except JWTError:
+                return {}
+
+        return inner
+
+    return wrapper
+
+
+@encode_jwt()
+def get_token(data: dict) -> dict | str:
+    return data
+
+
+@decode_jwt()
+def get_jwt(data: str) -> str | dict:
+    return data
+
+
+@router.get(
+    path='/check',
+    summary="Check access token",
+    description="Check access token"
+)
+async def check_token(token: str) -> dict:
+    return {'is_valid': 'login' in (get_jwt(token).keys())}
 
 
 @router.post(
@@ -25,10 +83,32 @@ async def sign_up(
         first_name=first_name,
         last_name=last_name
     )
-    return JSONResponse(
-        status_code=response['status_code'],
-        content=response['content']
-    )
+    if response['status_code'] == 201:
+        uuid = response['content']['uuid']
+        access_token = response['content']['access_token']
+        refresh_token = response['content']['refresh_token']
+
+        json_response = JSONResponse(
+            status_code=response['status_code'],
+            content={'uuid': uuid, "refresh_token": refresh_token, "token_type": 'bearer'}
+        )
+        json_response.set_cookie(
+            key="access_token",
+            value=access_token,
+            httponly=True
+        )
+        json_response.set_cookie(
+            key="refresh_token",
+            value=refresh_token,
+            httponly=True,
+            expires=10
+        )
+        return json_response
+    else:
+        return JSONResponse(
+            status_code=response['status_code'],
+            content=response['content']
+        )
 
 
 @router.get(
@@ -38,23 +118,27 @@ async def sign_up(
     description="Get one user with current uuid if exists"
 )
 async def get_user_by_uuid(
-        uuid: str,
-        service: UserService = Depends(get_user_service)
+       uuid: str,
+       access_token: str = Cookie(None),
+       service: UserService = Depends(get_user_service)
 ) -> UserResult | Response:
+    if not check_token(access_token):
+        return JSONResponse(
+            status_code=401,
+            content='Invalid access token'
+        )
     response: dict = await service.get_user_by_uuid(uuid)
     if response['status_code'] == 200:
         return UserResult(
             uuid=str(response['content']['uuid']),
             login=response['content']['login'],
             first_name=response['content']['first_name'],
-            last_name=response['content']['last_name'],
-            is_verified=response['content']['is_verified']
+            last_name=response['content']['last_name']
         )
     else:
         return JSONResponse(
-                status_code=response['status_code'],
-                content=response['content']
-            )
+            status_code=response['status_code'],
+            content=response['content'])
 
 
 @router.post(
@@ -64,8 +148,14 @@ async def get_user_by_uuid(
 )
 async def delete_user(
         uuid: str,
+        access_token: str = Cookie(None),
         service: UserService = Depends(get_user_service)
 ) -> Response:
+    if not check_token(access_token):
+        return JSONResponse(
+            status_code=401,
+            content='Invalid access token'
+        )
     response: dict = await service.remove_account(uuid)
     return JSONResponse(
         status_code=response['status_code'],
@@ -79,15 +169,36 @@ async def delete_user(
     description="Login by login and password"
 )
 async def login_user(
-        login: str,
-        password: str,
-        service: UserService = Depends(get_user_service)
+       login: str,
+       password: str,
+       service: UserService = Depends(get_user_service)
 ) -> Response:
     response: dict = await service.authenticate(login, password)
-    return JSONResponse(
-        status_code=response['status_code'],
-        content=response['content']
-    )
+    if response['status_code'] == 200:
+        access_token = response['content']['access_token']
+        refresh_token = response['content']['refresh_token']
+
+        json_response = JSONResponse(
+            status_code=response['status_code'],
+            content={"refresh_token": refresh_token, "token_type": 'bearer'}
+        )
+        json_response.set_cookie(
+            key="access_token",
+            value=access_token,
+            httponly=True
+        )
+        json_response.set_cookie(
+            key="refresh_token",
+            value=refresh_token,
+            httponly=True,
+            expires=10
+        )
+        return json_response
+    else:
+        return JSONResponse(
+            status_code=response['status_code'],
+            content=response['content']
+        )
 
 
 @router.post(
@@ -96,12 +207,18 @@ async def login_user(
     description="Update profile data except password"
 )
 async def update_user(
-        uuid: str,  # передается, чтобы можно было поменять логин
-        login: str,
-        first_name: str,
-        last_name: str,
-        service: UserService = Depends(get_user_service)
+       uuid: str,
+       login: str,
+       first_name: str,
+       last_name: str,
+       access_token: str = Cookie(None),
+       service: UserService = Depends(get_user_service)
 ) -> Response:
+    if not check_token(access_token):
+        return JSONResponse(
+            status_code=401,
+            content='Invalid access token'
+        )
     response: dict = await service.update_profile(uuid, login, first_name, last_name)
     return JSONResponse(
         status_code=response['status_code'],
